requirements.txt
asgiref==3.8.1
Django==5.2.3
django-filter==25.1
django-taggit==6.1.0
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
